// web/lib/api/generated/model/returnRequestsListStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

export type ReturnRequestsListStatus = typeof ReturnRequestsListStatus[keyof typeof ReturnRequestsListStatus];


export const ReturnRequestsListStatus = {
  APPROVED: 'APPROVED',
  CANCELLED: 'CANCELLED',
  REJECTED: 'REJECTED',
  REQUESTED: 'REQUESTED',
} as const;